view matrix working fine i guess

+++ blueprints/frontend/__init__.py
@@ -70,11 +70,11 @@ def fun(s):
             return int(s)
     except Exception:
         return None
-@frontend.route('donations', methods=['GET','POST'])
+@frontend.route('donations/', methods=['GET','POST'])
 @login_required
 def view_donations():
     #email = 'arunpandianp@gmail.com'
-    email = current_user.email
+    email = current_user.email_address
     if request.method == 'GET':
         donor = db.session.query(Donor).filter_by(email_address = email).first()
         if donor.is_admin:
@@ -83,7 +83,9 @@ def view_donations():
             donations = donor.donations
         return render_template('view_donations.htm', donations=donations)
     elif request.method == 'POST':
-        print request.args
+        isemail = False
+        if request.args.get('email'):
+            isemail = True
         donation_ids = map(lambda x : fun(x),request.form.keys())
         print donation_ids
         donor = db.session.query(Donor).filter_by(email_address = email).first()
@@ -99,15 +101,30 @@ def view_donations():
             donation_ids = donor_accessible_donation_ids
             donations=Donation.query.filter(Donation.id.in_(donation_ids)).all()
             print donations
-            return generate_zipped_receipts(donor, donations)
+            return generate_zipped_receipts(donor, donations,isemail)
+donation_made_text = '''Dear Donor,
 
-def generate_zipped_receipts(donor, donations):
+  Thank you for making a difference by donating for Team Everest.  Please find your E-receips tattached with this email for your donation.
+
+  Keep making a difference.
+
+  Thanks & Regards,
+  Team Everest
+  www.teameverestindia.org
+'''
+
+
+def generate_zipped_receipts(donor, donations,isemail):
     pdfFiles = []
     for donation in donations:
         data = {'Amount':donation.amount, 'ReceiptNo':donation.id, 'DonationDate': donation.date, 'OrgName': "Team Everest", 'OrgAddress': "chennai", 'DonationMode': donation.mode, 'DonorName':donor.name, 'DonorAddress' : donor.address, 'Certification_Number' : "213213dsfdaf3", "WordAmount":num2words(donation.amount) }
         pdf = makePdf(data)
         pdfFiles.append(pdf)
     s = zipFiles(pdfFiles)
+    if(isemail):
+        file = open(s)
+        sendEmail(donor.email_address, donation_made_text, file)
+        return redirect(request.referrer) 
     response_body = open(s).read()
     os.remove(s)
     headers = {"Content-Disposition": "attachment; filename=receipt.zip"}
